Fixed bug with Colorscheme Menu
Fixed a bug, which allowed the Options in the Colorscheme Menu to be clicked even though the Menu was closed.

Caused by opening the Settings Menu followed by the Colors Menu. If the Settings Menu is closed without closing the Colors Menu the state of the Colors Menu remains open and therefore set to visible.

Fixed by moving the State, which governs the Color Menus open state to the parent and setting it to false everytime the Settings Menu Button is clicked.

Considering to create a new Context just for the state of the menus in the future.