Styling
Gave Section Headtext (h2) a margin top of 8. When using the navigation Links to get to the Sections the Headtext isnt overlapped by the Header anymore.

Disabled outline for the links inside the ProjectCard Component. When the links are focused the img elements inside the links receive a ring around them.

Changed order of the links in the JSX. Livelink before the githublink.

Made the Contact Section unfocusable through tab by giving it a tabIndex of -1.