Finished Project Rework
Added last remaining styles to index.css for all the different colorschemes.

Added max-width for ProjectCard for responsiveness.

In the Projects Component projectsRef references the parent section element now instead of the div containing the Projectcards. This allows the hover effect to show effect while hovering over the Headtext aswell.